Updated ts samples as per reviews
